Extracted `HealthOperation`
Just a small refactor to make `InternalAPI` smaller, which will be expanded in future PRs.
